+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.4.0] - 2020.mm.dd
+
+**breaking changes in this version!**
+
+- a validation `type: :struct` that had been deprecated since ver. 1.2.2 was finally removed
+- the `func` check's callback arguments now aligned with `coerce_with` callback, they are: a parameter's name/value tuple (the first), all parameters map given to an operation (the second), the output of the validation callback fuction has been updated as well (check README for details)
+- `Exop.Chain`'s `operation` (`step`) now can be conditional with `if: _your_condition_func/1` option provided (see README for the details)
+- `Exop.Chain`'s `operation` now can coerce incoming parameters with `coerce_with: your_coerce_func/1` option before any further checks/invocations
+
 ## [1.3.5] - 2019.12.27
 
 - basic validation message extension with "got:" (credits to https://github.com/sgilson)